Rename handlers namespace to be more consistent with its functionality
